Continued progress toward modularization.
Pretty much have converted the old fwyplot.js and tmcpe.js mishmash
into a set of dojo widgets that implement components (TimeSpaceDiagram
and TestbedMap) and top-level application behavior (IncidentList and
IncidentView).

Modifed the grails tmcpe tag lib to push tmcpe onto the dojo import
path (so the components can be 'required'), and changed the relevant
.gsp files to use this approach.

There are some interface changes in there too---particularly the code
to show when the map is loading data (this is still a work in
progress).

Also some random changes to the vds JSON converter to
include the vds location.

	IncidentList.js