Beginnings of a working CAS implementation.
The approach taken here creates a custom CASUserDetails and CASUserDetailsService 
that overrides the default GormUserDetailsService (loaded in resources.groovy) to bypass looking 
up user data in local database tables in favor of pulling data directly from the CAS response or 
from an LDAP lookup.

The basics work, but it's more of a proof of concept than a long term solution.  Further work is necessary.