Got the basics of spring security working with CAS backed by LDAP for user details.
This approach uses the grails-spring-security-cas plugin to defer authentication to CAS.  This, by overriding the default implementation in resources.groovy, we set the UserDetailsService to use LDAP to obtain user information---including group affliations, which are translated into ROLEs using the CtmlabsLdapAuthoritiesGot the basics of spring security working with CAS backed by LDAP for user details.  Refs #194.