Added ability to download an xls workbook contining raw and processed incident data.  Refs #195.
Extended the IncidentFacilityImpactAnalysisController to accept
requests with the .xls extension to download.  E.g.,

http://localhost/tmcpe/incidentFacilityImpactAnalysis/show.xls?id=123

Will return an xls spreadsheet containing the raw and processed data
associated with facilityimpactanalysis number 123.

A direct link to this was added to the custom incident view (under the
time-space diagram).

The implementation uses apache's poi library to generate the workbook
on the fly.