Fixed buggy time and direction calcs in the DelayComputation module.
Incorrect computations on the time stamps meant that delays wouldn't
be calculated for incidents whose start/end times didn't fall on 5
minute boundaries.  Fixed this by correcting flawed modulus logic.

Also, it appears that the ordering on the W/E logic was reversed for getting stations upstream of an incident location.