Fixed bug in delay computation formulation.  Added commentary.  Added a "FORCE" equation.
EQ 3 was formulated to ensure that:
   "if j1,m1 is a boundary in space (-, downstream) at time m1, the sum of all D's later that M1 at section J1 must be <= 0"

This constraint is unnecessary and it seems to overconstrain the
problem (as formulated).  Removing it produces correct solutions
(based upon visual inspection of a sample.

The FORCEEQ can be used to require certain cells are in the solution.
It is primarily a debugging tool.  It's use is strictly through
hardcode modification, though a command-line version is possible.