Fixed some bugs in the delay formulation.  Added more options for tuning as well.
The delay formulation was incorrectly constraining the solution based
upon boundaries.  Basically, the problem was a summation going in the
wrong direction (upstream instead of downstream).  The solution
quality is much better now.

Also added the capability of limiting the (implied) speed of
shockwaves in the solution by adding additional constraints (eqs8).

All of these tweaks are tunable from command line parameters.