Further tuning of the delay calculation constraints.
Tweaked the outer rectangle constriants to be blanket constraints at the upstream/later edge of the incident region.  Basically, if a cell is at the upstream/later edge of the incident region then there can't be any more incident cells later in /any/ section.

Also added the ability to force the value of specific cells from the command line.  This is a debugging tool.

Made clearing and loading shockwave constraint parameters independent.