Fixed directional bugs related to formulating the delay computation.  Added additional constraints.
The existing code was not correctly ordering the GAMS program for all
freeway directions.  Similfied the sorting code and synchronized with
GUI logic.

The ability to limit the program require that the incident boundary
start within a particular window is also now included as an option.

Fixed some logic problems in the LocationParser code that selects the
best vds location based upon distance from the icad location.
Existing code was sorting furthest to nearest rather than nearest to
furthest.