Added support for caching pems averages.  Fixes #306.
Making queries to the database to obtain statistics is expensive.  The
changes in this commit make it so that these computations only have to
happen once---on demand---at which point they are added to the
pems_5min_annual_avg table.  Future need for that data (e.g., if an
analysis is re-run with different parameters) will use the data from
that table instead of recomputing them.

The use of the cache can be turned off with the switch
      --dc-dont-use-pems-cache

As part of this change, the query for average data has been separated
such that the average for each period is queried separately.  Probably
marginally slower but more flexible.