Added option to use osm data to obtain approximate location of incidents.
Fixes #324.

Often, the facility and cross street in the activity logs for incident
location do not readily match to a vds name.  In the absense of icad
x/y data, this leaves a lot of incident unlocated, and therefore
unanalyzed.  The osm highway data provides a means for obtaining an
approximate location by looking for the intersection of a mainline
facility with other ways in the osm data.  Since this is a large set,
we cull it by only looking at ways whose names are sufficiently
similar to the provided cross street.  Given this hint, the
LocationParser is pretty good at getting close to the actual incident
location.

Also,

* added some improvements to the test scripts for the LocationParser.
* added DBIx::Class schemas for the OSM database.

Augmented the osm database as follows (see osm-routes.sql)

* loaded the trigram support (from postgres/contrib)
* added a routes view for convenience.