Added date to sort field of the write_to_db routine in the delay computation.
The prior approach sorted results on time-of-day only which produced a bad time
offset when the incident crossed the midnight boundary.  By sorting on date + '
' + time, this problem is resolved and the offset is correst.  Also added the
timestamp to the analyzed_timesteps in the incident facility impact analysis
sent to the browser.

   * Fixes #389
   * Fixes #390