Generalized handling of listGroup parameters.
The problem here is that to create a good query interface, we need to be able to
present the user with the set of possible group/stack/filter options.  The
existing implementation hardcoded these in the controller logic in a manner that
would be difficult to maintain.  This commit implements an approach that maps
possible parameters to the backend database implementations.  This mapping
permits the possible parameters to be offered to the UI via the controller and
interpreted by the controller when sent as part of a GET.

* Refs #863